fix trainer confirmation
